pyproject.toml
[tool.poetry]
name = "apibility"
version = "0.1.0"
description = ""
authors = ["Randi Guarcas <rguarcas18@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11,<3.12"
fastapi = "^0.115.2"
uvicorn = "^0.31.1"
llama-index = "^0.11.17"
python-dotenv = "^1.0.1"
llama-index-llms-ollama = "^0.3.4"
llama-index-embeddings-huggingface = "^0.3.1"
llama-index-vector-stores-chroma = "^0.2.1"
llama-index-llms-groq = "^0.2.0"
llama-index-callbacks-arize-phoenix = "^0.2.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
